third_party/blink/web_tests/platform/generic/http/tests/security/opened-document-security-origin-resets-on-navigation-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
PASS successfullyParsed is true

TEST COMPLETE
